% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2g.R
\name{beta2es}
\alias{beta2es}
\title{Conversion from a Beta coefficient to an effect size object}
\usage{
beta2es(Study, Type, Beta, SDy, N1, N2, Y, X, Ryy, Rxx)
}
\description{
This function converts form a beta coefficient to an effect size object
}
