% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2g.R
\name{f2es}
\alias{f2es}
\title{Conversion from a F coefficient to an effect size object}
\usage{
f2es(study, type, f, n1, n2, X, Y, Rxx = 0, Ryy = 0)
}
\description{
This function converts form a F coefficient to an effect size object
}
