% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mat2DF.R
\name{Mat2DF}
\alias{Mat2DF}
\title{Conversion from a correlation matrix to an effect size object}
\usage{
Mat2DF(Study, Mat.Cor, Vars, n.X, s.Y, N, low.diag = T, diag.val = T)
}
\description{
This function converts form a correlation matrix to an effect size object
}
