% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanSE2g.R
\name{mean.se2es}
\alias{mean.se2es}
\title{Conversion from a mean and standard error to an effect size object}
\usage{
\method{mean}{se2es}(study, type, m1, se1, n1, m2, se2, n2, X, Y, Rxx = 0, Ryy = 0)
}
\description{
This function converts form a mean and standard error to an effect size object
}
