% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2g.R
\name{beta2es}
\alias{beta2es}
\title{Conversion from a Beta coefficient to an effect size object}
\usage{
beta2es(study, type, beta, sd, n1, n2, X, Y, Rxx = 0, Ryy = 0)
}
\description{
This function converts form a beta coefficient to an effect size object
}
